% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RNAseq_stats.R
\name{count_kmer_in_fastq}
\alias{count_kmer_in_fastq}
\title{Title of the Function}
\usage{
count_kmer_in_fastq(fq, specific_kmer)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of what the function returns
}
\description{
Description of what the function does.
}
