% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_virus}
\alias{download_virus}
\title{Title: Function to download the virus genome(s).}
\usage{
download_virus(path_csv_file, path_output, virus_name = "all")
}
\arguments{
\item{path_csv_file}{path to the csv file containing a column "name" and a column "accession_code" for viruses of interest.}

\item{path_output}{path to the directory where the genomes will be stored (one .fasta file will be generated by unique name in the column name of the virus_list).}

\item{virus_name}{if "all" then all viruses will be downloaded from the virus_list, otherwise the specific virus (1 max) will be downloaded}
}
\value{
Saves the .fasta files inside the "path_output"
}
\description{
Description: This function will get the genomes of viruses from NCBI using their accession codes. These accession codes are stored inside a csv file called (Virus_list) that contains a columns with the virus type and subtype as well as one or multiple accession code(s) per virus.
This function will then store all viruses genomes in a fasta according to their name. If multiple lines have the same name in the virus_list they will be combined into one fasta file and considered as variation of the virus genome.
}
