% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_dist_df.R
\name{gen_kmers_pos_df}
\alias{gen_kmers_pos_df}
\title{Title of the Function}
\usage{
gen_kmers_pos_df(fasta_file, kmers_list, virus_type)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of what the function returns
}
\description{
Description of what the function does.
}
