% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_list.R
\name{gen_obs_kmers}
\alias{gen_obs_kmers}
\title{Title of the Function}
\usage{
gen_obs_kmers(fasta_file, k)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of what the function returns
}
\description{
Description of what the function does.
}
