% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats_kmers}
\alias{get_stats_kmers}
\title{Title of the Function}
\usage{
get_stats_kmers(df_virus)
}
\arguments{
\item{x}{Description of parameter x}
}
\value{
Description of what the function returns
}
\description{
Description of what the function does.
}
